+++ Services/Commands/Tools/TokenInfoApi.cs
@@ -5,10 +5,10 @@ namespace Telegram.CryptoTracker.Bot.Services.Commands.Tools
     public abstract class TokenInfoApi
     {
 
-        public abstract Task<double> GetPriceTokenFromApi();
+        public abstract Task<decimal> GetPriceTokenFromApi();
 
 
-        public async Task<double> GetPriceToken()
+        public async Task<decimal> GetPriceToken()
         {
            return await GetPriceTokenFromApi();
         }
