+++ Services/MySQL/DataDB.cs
@@ -7,9 +7,9 @@
         public long Chat_id { get; set; }
         public string Token { get; set; }
 
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
 
-        public double PriceAverage { get; set; }
+        public decimal PriceAverage { get; set; }
 
     }
 }
